+++ packages/core/src/controllers/AccountController.ts
@@ -1,5 +1,10 @@
 import { CoreHelperUtil } from '../utils/CoreHelperUtil.js'
-import type { CaipAddress, ConnectedWalletInfo, SocialProvider } from '../utils/TypeUtil.js'
+import type {
+  AccountType,
+  CaipAddress,
+  ConnectedWalletInfo,
+  SocialProvider
+} from '../utils/TypeUtil.js'
 import type { Balance } from '@web3modal/common'
 import { BlockchainApiController } from './BlockchainApiController.js'
 import { SnackController } from './SnackController.js'
@@ -9,7 +14,7 @@ import type { W3mFrameTypes } from '@web3modal/wallet'
 import { ChainController } from './ChainController.js'
 import type { Chain } from '@web3modal/common'
 import { NetworkController } from './NetworkController.js'
-import { proxy } from 'valtio'
+import { proxy, ref } from 'valtio'
 
 // -- Types --------------------------------------------- //
 export interface AccountControllerState {
@@ -17,6 +22,8 @@ export interface AccountControllerState {
   currentTab: number
   caipAddress?: CaipAddress
   address?: string
+  addressLabels: Map<string, string>
+  allAccounts: AccountType[]
   balance?: string
   balanceSymbol?: string
   profileName?: string | null
@@ -25,6 +32,7 @@ export interface AccountControllerState {
   smartAccountDeployed?: boolean
   socialProvider?: SocialProvider
   tokenBalance?: Balance[]
+  shouldUpdateToAddress?: string
   connectedWalletInfo?: ConnectedWalletInfo
   preferredAccountType?: W3mFrameTypes.AccountType
   socialWindow?: Window
@@ -35,7 +43,9 @@ const state = proxy<AccountControllerState>({
   isConnected: false,
   currentTab: 0,
   tokenBalance: [],
-  smartAccountDeployed: false
+  smartAccountDeployed: false,
+  addressLabels: new Map(),
+  allAccounts: []
 })
 
 // -- Controller ---------------------------------------- //
@@ -118,6 +128,25 @@ export const AccountController = {
       ChainController.setAccountProp('tokenBalance', tokenBalance, chain)
     }
   },
+  setShouldUpdateToAddress(address: string) {
+    ChainController.setAccountProp('shouldUpdateToAddress', address)
+  },
+
+  setAllAccounts(accounts: AccountType[], chain?: Chain) {
+    ChainController.setAccountProp('allAccounts', accounts, chain)
+  },
+
+  addAddressLabel(address: string, label: string) {
+    const map = ChainController.getAccountProp('addressLabels') || new Map()
+    map.set(address, label)
+    ChainController.setAccountProp('addressLabels', map)
+  },
+
+  removeAddressLabel(address: string) {
+    const map = ChainController.getAccountProp('addressLabels') || new Map()
+    map.delete(address)
+    ChainController.setAccountProp('addressLabels', map)
+  },
 
   setConnectedWalletInfo(
     connectedWalletInfo: AccountControllerState['connectedWalletInfo'],
@@ -141,7 +170,7 @@ export const AccountController = {
 
   setSocialWindow(socialWindow: AccountControllerState['socialWindow'], chain?: Chain) {
     if (socialWindow) {
-      ChainController.setAccountProp('socialWindow', socialWindow, chain)
+      ChainController.setAccountProp('socialWindow', ref(socialWindow), chain)
     }
   },