+++ packages/wagmi/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @web3modal/wagmi
 
+## 5.0.6
+
+### Patch Changes
+
+- fix: Social Login illegal invocation issue. Wagmi tests
+
+- Updated dependencies []:
+  - @web3modal/polyfills@5.0.6
+  - @web3modal/scaffold@5.0.6
+  - @web3modal/scaffold-react@5.0.6
+  - @web3modal/scaffold-utils@5.0.6
+  - @web3modal/scaffold-vue@5.0.6
+  - @web3modal/siwe@5.0.6
+
+## 5.0.5
+
+### Patch Changes
+
+- feat: universal link internal flag. Add kotlin assetlinks. Fix email truncation'
+
+- Updated dependencies []:
+  - @web3modal/polyfills@5.0.5
+  - @web3modal/scaffold@5.0.5
+  - @web3modal/scaffold-react@5.0.5
+  - @web3modal/scaffold-utils@5.0.5
+  - @web3modal/scaffold-vue@5.0.5
+  - @web3modal/siwe@5.0.5
+
 ## 5.0.4
 
 ### Patch Changes